make edsc methods modular
* export as `dataHelper` instead of `L`
