+++ web/js/edsc.js
@@ -1,9 +1,14 @@
+import edscArc from './edsc/arc';
+import edscCoordinate from './edsc/coordinate';
+import edscGeoUtil from './edsc/geoUtil';
+import edscInterpolation from './edsc/interpolation';
+import edscSphericalPolygon from './edsc/spherical-polygon';
 var edsc = {};
 edsc.map = {};
+edsc.config = null;
 
 var L = {};
-edsc.map.L = L;
-
+edsc.L = L;
 L.Polyline = {
   prototype: {}
 };
@@ -50,21 +55,25 @@ L.Util = {
     return value.constructor === Array;
   }
 };
-
-(function () {
-  function LatLng (lat, lng) {
-    if (lat.lat) {
-      var latlng = lat;
-      this.lat = latlng.lat;
-      this.lng = latlng.lng;
-    } else {
-      this.lat = lat;
-      this.lng = lng;
-    }
+function LatLng (lat, lng) {
+  if (lat.lat) {
+    var latlng = lat;
+    this.lat = latlng.lat;
+    this.lng = latlng.lng;
+  } else {
+    this.lat = lat;
+    this.lng = lng;
   }
+}
 
-  L.LatLng = LatLng;
-  L.latLng = function (lat, lng) {
-    return new L.LatLng(lat, lng);
-  };
-})();
+L.LatLng = LatLng;
+L.latLng = function (lat, lng) {
+  return new L.LatLng(lat, lng);
+};
+edsc.Coordinate = edscCoordinate(L);
+edsc.Arc = edscArc(L, edsc.Coordinate);
+edsc.geoutil = edscGeoUtil(L, edsc.Coordinate, edsc.Arc, edsc.config);
+edsc.interpolation = edscInterpolation(L, edsc.geoutil.gcInterpolate);
+edsc.L.sphericalPolygon = edscSphericalPolygon(L, edsc.geoutil, edsc.Arc, edsc.Coordinate);
+
+export {L as dataHelper};
